move suggest items to folder
