+++ lib/src/types/suggest/suggest_item.dart
