+++ lib/src/types/suggest/suggest_type.dart
