+++ lib/yandex_mapkit.dart
@@ -27,9 +27,9 @@ part 'src/types/polyline.dart';
 part 'src/types/polyline_style.dart';
 part 'src/types/rotation_type.dart';
 part 'src/types/screen_point.dart';
-part 'src/types/suggest_item.dart';
-part 'src/types/suggest_item_type.dart';
-part 'src/types/suggest_type.dart';
+part 'src/types/suggest/suggest_item.dart';
+part 'src/types/suggest/suggest_item_type.dart';
+part 'src/types/suggest/suggest_type.dart';
 part 'src/types/driving/driving_route.dart';
 part 'src/types/driving/request_point.dart';
 part 'src/types/driving/driving_session.dart';
